--- packages/hoprd/hopr-admin/fetch/aliases.js
@@ -1,7 +1,0 @@
-import { getReq, postReq } from './client'
-
-export const getAliases = () => getReq("aliases")
-
-export const setAliases = (peerId, alias) => {
-  postReq("aliases/", {"peerId": peerId, "alias": alias})
-}
