--- packages/hoprd/hopr-admin/fetch/node.js
@@ -1,7 +1,0 @@
-import { getReq, postReq } from './client'
-
-export const getNodeInfo = () => getReq("node/info")
-
-export const getNodeVer = () => getReq("node/version")
-
-export const pingNodePeer = (peerId) => postReq("node/ping", {peerId: peerId})
