--- packages/hoprd/hopr-admin/fetch/settings.js
@@ -1,3 +1,0 @@
-import { getReq } from './client'
-
-export const getSettings = () => getReq("settings")
